chore: remove duplicated account service
